Fixes AIX test problems Disables on hpux two wsman tests that have issues on hp
